break out Carosel component
