+++ frontend/packages/data-portal/app/components/MLChallenge/CompletedMLChallenge/components/OtherWinners/components/Carousel.tsx
@@ -1,0 +1,128 @@
+import { Button, IconNameToSizes } from '@czi-sds/components'
+import { ReactElement, useState } from 'react'
+
+import { cns } from 'app/utils/cns'
+
+export function Carousel({
+  totalCards,
+  leftIcon,
+  rightIcon,
+  gap,
+  numberOfVisibleCards,
+  totalWidth,
+  children,
+}: {
+  totalCards: number
+  leftIcon?: keyof IconNameToSizes | ReactElement
+  rightIcon?: keyof IconNameToSizes | ReactElement
+  gap: number
+  numberOfVisibleCards: number
+  totalWidth: number
+  children: Array<ReactElement>
+}) {
+  const [carouselPosition, setCarouselPosition] = useState<number>(1)
+  const incompleteLastSlideCardNum = totalCards % numberOfVisibleCards
+  const numberOfSlides = Math.ceil(totalCards / numberOfVisibleCards)
+  const numberOfSlidesArray = [...Array(numberOfSlides).keys()]
+  const cardWidth =
+    (totalWidth - (numberOfVisibleCards - 1) * gap) / numberOfVisibleCards
+
+  const getCarouselOffset = ({ position }: { position: number }) => {
+    // if it is not the last slide or if the last slide has the full amount of visible cards
+    const getPreviousSlidesWidth = (includeLastSlide: boolean) => {
+      const totalSlidesWidth =
+        totalWidth * (position - (includeLastSlide ? 1 : 2)) +
+        gap * (position - (includeLastSlide ? 1 : 2))
+      // We add an extra gap to center the carousel within the viewport after including box shadows
+      return totalSlidesWidth
+    }
+
+    if (position !== numberOfSlides || incompleteLastSlideCardNum === 0) {
+      return getPreviousSlidesWidth(true)
+    }
+    // If the last slide has less than the total possible numberOfVisibleCards
+    const offset = getPreviousSlidesWidth(false)
+    const offsetLastSlide = (cardWidth + gap) * incompleteLastSlideCardNum
+    return offset + offsetLastSlide
+  }
+
+  const changeCarouselPosition = (direction: 'left' | 'right') => {
+    if (direction === 'left') {
+      setCarouselPosition((prev) => {
+        if (prev === 1) {
+          return prev
+        }
+        return prev - 1
+      })
+    } else {
+      setCarouselPosition((prev) => {
+        if (prev === numberOfSlides) {
+          return prev
+        }
+        return prev + 1
+      })
+    }
+  }
+  return (
+    <div>
+      <div className="overflow-x-clip w-[calc(100%+40px)] translate-x-[-20px]">
+        <div
+          style={{
+            right: `${getCarouselOffset({ position: carouselPosition })}px`,
+            gridTemplateColumns: `repeat(${totalCards}, ${cardWidth}px)`,
+          }}
+          className={cns(
+            'grid gap-sds-xxl',
+            'translate-x-[20px] relative',
+            'transition-all duration-600 ease-in-out',
+          )}
+        >
+          {children}
+        </div>
+      </div>
+      <div className="flex justify-center mt-sds-xxl items-center gap-sds-xl">
+        <Button
+          sdsType="primary"
+          sdsStyle="icon"
+          icon={leftIcon ?? 'ChevronLeft'}
+          sdsSize="small"
+          onClick={() => changeCarouselPosition('left')}
+          disabled={carouselPosition === 1}
+        />
+        <div
+          style={{
+            gridTemplateColumns: `repeat(${numberOfSlides}, ${
+              numberOfSlides > 4 ? 20 : 45
+            }px)`,
+          }}
+          className={`grid grid-cols-${numberOfSlides} gap-sds-s grid-rows-[4px]`}
+        >
+          {numberOfSlidesArray.map((_, index) => (
+            <div
+              // eslint-disable-next-line react/no-array-index-key
+              key={index}
+              className={`
+                  rounded-full
+                  h-full
+                  bg-sds-color-primitive-gray-200
+                  ${
+                    index === carouselPosition - 1
+                      ? 'bg-sds-color-semantic-component-accent-icon'
+                      : 'bg-sds-color-primitive-gray-200'
+                  }
+                `}
+            />
+          ))}
+        </div>
+        <Button
+          sdsType="primary"
+          sdsStyle="icon"
+          icon={rightIcon ?? 'ChevronRight'}
+          sdsSize="small"
+          onClick={() => changeCarouselPosition('right')}
+          disabled={carouselPosition === numberOfSlides}
+        />
+      </div>
+    </div>
+  )
+}
